+++ test/logger/mocks/logger_mock.h
@@ -14,7 +14,8 @@ class LoggerMock : public ILogger {
     MOCK_METHOD2(logInstructionCount, void(PlayerId player_id, size_t count));
     MOCK_METHOD3(logError, void(state::PlayerId player_id, ErrorType error_type,
                                 std::string message));
-    MOCK_METHOD2(logFinalGameParams, void(state::PlayerId player_id,
-                                          std::array<size_t, 2> final_scores));
+    MOCK_METHOD2(logFinalGameParams,
+                 void(state::PlayerId player_id,
+                      std::array<uint64_t, 2> final_scores));
     MOCK_METHOD1(writeGame, void(std::ostream &write_stream));
 };
